% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooni-package.R
\docType{package}
\name{ooni}
\alias{ooni}
\alias{ooni-package}
\title{Tools to Access the Open Observatory of Network Interference ('OONI')}
\description{
The 'Open Observatory of Network Interference' (\url{http://ooni.io/}) is
a free software, global observation network for detecting censorship,
surveillance and traffic manipulation on the internet. Tools are provided to
query the 'API' and download both bulk metadata and test observation data
files.
}
\details{
The OONI API is nice for cursory analysis or some integrations with other
systems (e.g. OONI Explorer relies on it), but it’s currently not possible to
run any dataset scan that runs for more than a minute using the OONI API. So
the API is not the best option if the query does heavy scanning of metadata.
Also, implementation of pagination in the OONI API next_url is far from perfect
and may fail with greater than zero offsets.
}
\author{
Bob Rudis (bob@rud.is)
}
