% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-measurement.R
\name{ooni_get_measurement}
\alias{ooni_get_measurement}
\title{Retrieve an OONI measurement}
\usage{
ooni_get_measurement(measurement_id)
}
\arguments{
\item{measurement_id}{the id of the measurement to retrieve}
}
\description{
Retrieve an OONI measurement
}
