% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files.R
\name{ooni_list_files}
\alias{ooni_list_files}
\title{List measurement files}
\usage{
ooni_list_files(probe_cc = NULL, probe_asn = NULL, test_name = NULL,
  since = NULL, until = NULL, since_index = NULL,
  order_by = c("test_start_time", "probe_cc", "report_id", "probe_asn",
  "test_name", "index", "idx", "report_no"), ascending = TRUE, offset = 0L,
  limit = 100L)
}
\arguments{
\item{probe_cc}{two letter country code}

\item{probe_asn}{autonomous system number in the format "ASXXX"}

\item{test_name}{name of the test}

\item{since, until}{start/end dates of when measurements were run. A \code{POSIXct} object,
\code{Date} object or character string that can be converted to one of those
objects.}

\item{since_index}{return results only strictly greater than the provided index}

\item{order_by}{how the results should be ordered by. One of "\code{test_start_time}",
"\code{probe_cc}", "\code{report_id}", "\code{probe_asn}", "\code{test_name}", "\code{index}",
"\code{idx}", or "\code{report_no}". Default is "\code{test_start_time}".}

\item{ascending}{should the order be ascending or descending}

\item{offset}{offset into the paginated result set (default: \code{0})}

\item{limit}{number of records to return (default & max are \code{100})}
}
\description{
List measurement files
}
