% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-measurements.R
\name{ooni_list_measurements}
\alias{ooni_list_measurements}
\title{List measurements}
\usage{
ooni_list_measurements(test_name = c("web_connectivity", "http_requests",
  "dns_consistency", "http_invalid_request_line", "bridge_reachability",
  "tcp_connect", "http_header_field_manipulation", "http_host",
  "multi_protocol_traceroute", "meek_fronted_requests_test", "whatsapp",
  "vanilla_tor", "facebook_messenger", "ndt", "dash", "telegram"),
  probe_cc = NULL, probe_asn = NULL, confirmed = TRUE, report_id = NULL,
  input = NULL, since = NULL, until = NULL, since_index = NULL,
  order_by = c("test_start_time", "probe_cc", "probe_asn", "test_name",
  "input"), ascending = TRUE, offset = 0L, limit = 100L)
}
\arguments{
\item{test_name}{the name of the measurement test. One of "\code{web_connectivity}",
"\code{http_requests}", "\code{dns_consistency}", "\code{http_invalid_request_line}",
"\code{bridge_reachability}", "\code{tcp_connect}", "\code{http_header_field_manipulation}",
"\code{http_host}", "\code{multi_protocol_traceroute}", "\code{meek_fronted_requests_test}",
"\code{whatsapp}", "\code{vanilla_tor}", "\code{facebook_messenger}", "\code{ndt}", "\code{dash}",
"\code{telegram}". Defaults to "\code{web_connectivity}".}

\item{probe_cc}{two letter country code}

\item{probe_asn}{autonomous system number in the format "ASXXX"}

\item{confirmed}{logical (optional); return only measurements that have confirmed network
anomalies? (i.e. we found a blockpage, a middlebox was found, the IM app is blocked,
etc.). Default is \code{TRUE} meaning return only confirmed anomalies?}

\item{report_id}{the report id to search measurements for}

\item{input}{URL or IP address to search measurements for}

\item{since, until}{start/end dates of when measurements were run. A \code{POSIXct} object,
\code{Date} object or character string that can be converted to one of those
objects.}

\item{since_index}{return results only strictly greater than the provided index}

\item{order_by}{how the results should be ordered by. One of "\code{test_start_time}",
"\code{probe_cc}", "\code{report_id}", "\code{probe_asn}", "\code{test_name}", "\code{index}",
"\code{idx}", or "\code{report_no}". Default is "\code{test_start_time}".}

\item{ascending}{should the order be ascending or descending}

\item{offset}{offset into the paginated result set (default: \code{0})}

\item{limit}{number of records to return (default & max are \code{100})}
}
\description{
List measurements
}
