% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next-page.R
\name{ooni_next_pg}
\alias{ooni_next_pg}
\title{Retrieve the next page of results from an OONI paginated API endpoint}
\usage{
ooni_next_pg(res)
}
\arguments{
\item{res}{return value from a paginated OONI API call}
}
\description{
Retrieve the next page of results from an OONI paginated API endpoint
}
\concept{pagination helpers}
