% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next-page.R
\name{ooni_pg_info}
\alias{ooni_pg_info}
\title{Retrieve pagination metadata for a current request result object}
\usage{
ooni_pg_info(res)
}
\arguments{
\item{res}{return value from a paginated OONI API call}
}
\value{
list containing the API method, total record count, current page,
page results limit, offset and total number of pages. If \code{count} or
\code{pages} is \code{-1} then there are a larger number of results and you
should consider using better search filters.
}
\description{
Retrieve pagination metadata for a current request result object
}
