% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooni-package.R
\docType{package}
\name{ooni}
\alias{ooni}
\alias{ooni-package}
\title{Tools to Access the Open Observatory of Network Interference ('OONI')}
\description{
The 'Open Observatory of Network Interference' (\url{http://ooni.io/}) is
a free software, global observation network for detecting censorship,
surveillance and traffic manipulation on the internet. Tools are provided to
query the 'API' and download both bulk metadata and test observation data
files.
}
\author{
Bob Rudis (bob@rud.is)
}
